Refactor utilities to raise exceptions
